+++ changelog.md
@@ -7,6 +7,7 @@ FORMAT: 1A
 * DOC: API Documentation now denotes length, minvalue, maxlength, maxvalue.
 * FIX: Permit contentType to be specified during update without a content stream.
 * CFG: Database tool compatibility with MySQL 5.7 (requires parameter show_compatibility_56 = 1)
+* FIX: Add support for creating objects owned by group with pathing at time of creation.
 * FIX: Security fix for client certificate checks to AAC and peer odrives
 
 ## Release v1.0.1.21