+++ changelog.md
@@ -10,6 +10,7 @@ FORMAT: 1A
 * FIX: Add support for creating objects owned by group with pathing at time of creation.
 * FIX: Security fix for client certificate checks to AAC and peer odrives
 * DOC: API Documentation correction to create object size sample denoting contentSize as a number.
+* FIX: Build process will install graphviz to satisfy plantuml need for dot for diagram generation.
 
 ## Release v1.0.1.21
 --------------------